:recycle: refactor code and add index file
